dex2oat: Use CLOCK_MONOTONIC, not REALTIME for a watchdog!
Prior to this commit, dex2oat was using the realtime clock to determine
whether or not to kill dex2oat when it ran too long.

This causes problems when the time on the system changes during dex2oat.

In particular, on victara, the time daemon would start up during dex2oat
of boot.art which causes that critical dex2oat operation to fail on boot.

**Fixes lag on moto/msm8916 devices

Change-Id: I8be1e09479006bf0692bb9b29f3eab2d140096bb